WidgetsServer/requirements.txt
grpcio
grpcio-tools
wxPython
